added combined probability, but it suffers from floating point underflow

+++ src/BayesianClassifier/DataSet.cs
@@ -145,7 +145,7 @@ namespace BayesianClassifier
             var totalCount = _setSize; // TODO: cache inverse set size
             var vocabularySize = TokenCount;
 
-            return (tokenCount + alpha)/(totalCount + alpha*vocabularySize);
+            return (double)(tokenCount + alpha)/(double)(totalCount + alpha*vocabularySize);
         }
 
         /// <summary>
